sentiment-analyzer/requirements.txt
Flask
python-dotenv
google-generativeai
Jinja2
Werkzeug